tests/node-traversal.rs: move tests related to `NodeRef::find_descdendants` to src/mini_selector/extension.rs
