+++ src/mini_selector/extension.rs
@@ -120,3 +120,66 @@ impl NodeRef<'_> {
         Ok(res)
     }
 }
+
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{Document, NodeId};
+
+    #[test]
+    fn test_names() {
+        let sel = r#"body td a"#;
+        let parsed = parse_selector_list(sel).unwrap();
+        assert_eq!(parsed.1.len(), 3);
+    }
+
+    #[cfg_attr(not(target_arch = "wasm32"), test)]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn test_node_find_descendants() {
+        let html_contents = include_str!("../../test-pages/hacker_news.html");
+        let doc = Document::from(html_contents);
+        let a_sel = doc.select(r#"body td.title a[href^="https://"]"#);
+        let expected_ids: Vec<NodeId> = a_sel.nodes().iter().map(|n| n.id).collect();
+    
+        let root = doc.root();
+        let got_ids: Vec<NodeId> = root
+            .find_descendants(r#"body td.title a[href^="https://"]"#)
+            .iter()
+            .map(|n| n.id)
+            .collect();
+    
+        assert_eq!(got_ids, expected_ids);
+    
+        let a_sel = doc.select("a");
+        let expected_ids: Vec<NodeId> = a_sel.nodes().iter().map(|n| n.id).collect();
+        let got_ids_a: Vec<NodeId> = root.find_descendants("a").iter().map(|n| n.id).collect();
+        assert_eq!(got_ids_a, expected_ids);
+    
+        let len_fin_ne = root.find_descendants("body td p").len();
+        assert_eq!(len_fin_ne, 0);
+        let len_sel_ne = doc.select("body td p").length();
+        assert_eq!(len_sel_ne, 0)
+    }
+
+    #[cfg_attr(not(target_arch = "wasm32"), test)]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn test_node_find_descendant_combinators() {
+        let html_contents = include_str!("../../test-pages/hacker_news.html");
+        let doc = Document::from(html_contents);
+
+        let selectors = ["body td.title a", "body td.title > a"];
+
+        for sel in selectors {
+            let a_sel = doc.select(sel);
+            let expected_ids: Vec<NodeId> = a_sel.nodes().iter().map(|n| n.id).collect();
+
+            let root = doc.root();
+            let got_ids: Vec<NodeId> = root.find_descendants(sel).iter().map(|n| n.id).collect();
+
+            assert_eq!(got_ids, expected_ids);
+        }
+    }
+
+}
